Move taskList into its own screen
