+++ app/src/modules/selectedScreen/index.tsx
@@ -5,15 +5,17 @@ export const slice = createSlice({
   initialState: 'task',
   reducers: {
     openTask: () => 'task',
+    openTasksList: () => 'tasksList',
     openShortcuts: () => 'shortcuts',
     openChangelog: () => 'changelog',
   },
 });
 
-export const { openTask, openShortcuts, openChangelog } = slice.actions;
+export const {
+  openTask,
+  openTasksList,
+  openShortcuts,
+  openChangelog,
+} = slice.actions;
 
 export default slice.reducer;
-
-export function getSelectedScreen(store: { selectedScreen: string }) {
-  return store.selectedScreen;
-}
