+++ app/src/screens/index.tsx
@@ -1,13 +1,15 @@
 import * as React from 'react';
 import useSelector from 'hooks/useSelector';
-import { getSelectedScreen } from 'modules/selectedScreen';
+import { getSelectedScreen } from 'modules/selectors';
 
 import task from './task';
+import tasksList from './tasksList';
 import shortcuts from './shortcuts';
 import changelog from './changelog';
 
 const SCREENS = {
   task,
+  tasksList,
   shortcuts,
   changelog,
 };
