+++ app/src/screens/tasksList/styles.module.css
@@ -1,0 +1,7 @@
+.link {
+  cursor: pointer;
+}
+
+.link:hover {
+  text-decoration: underline;
+}
